+++ psm_utils/__init__.py
@@ -1,6 +1,6 @@
 """Common utilities for parsing and handling PSMs, and search engine results."""
 
-__version__ = "0.1.0"
+__version__ = "0.2.0"
 
 from psm_utils.peptidoform import Peptidoform
 from psm_utils.psm import PSM