+++ psm_utils/io/tsv.py
@@ -50,7 +50,7 @@ from __future__ import annotations
 import ast
 import csv
 from pathlib import Path
-from typing import Optional, Union
+from typing import Optional
 
 from psm_utils.io._base_classes import ReaderBase, WriterBase
 from psm_utils.io.exceptions import PSMUtilsIOException
@@ -60,9 +60,6 @@ from psm_utils.psm_list import PSMList
 
 class TSVReader(ReaderBase):
     """Reader for psm_utils TSV format."""
-
-    def __init__(self, filename: Union[str, Path], *args, **kwargs) -> None:
-        super().__init__(filename, *args, **kwargs)
 
     def __iter__(self):
         """Iterate over file and return PSMs one-by-one."""
@@ -92,11 +89,11 @@ class TSVReader(ReaderBase):
         rescoring_features = {}
         for k, v in entry.items():
             if k.startswith("provenance:"):
-                provenance_data[k[11:]] = v
+                provenance_data[k[11:]] = str(v)
             elif k.startswith("meta:"):
-                metadata[k[5:]] = v
+                metadata[k[5:]] = str(v)
             elif k.startswith("rescoring:"):
-                rescoring_features[k[10:]] = v
+                rescoring_features[k[10:]] = str(v)
             else:
                 parsed_entry[k] = v
 
@@ -116,7 +113,7 @@ class TSVWriter(WriterBase):
 
     def __init__(
         self,
-        filename: Union[str, Path],
+        filename: str | Path,
         example_psm: Optional[PSM] = None,
         *args,
         **kwargs,