+++ psm_utils/io/xtandem.py
@@ -45,7 +45,6 @@ Notes
 from __future__ import annotations
 
 from pathlib import Path
-from typing import Union
 
 import numpy as np
 from pyteomics import mass, tandem
@@ -58,7 +57,9 @@ from psm_utils.psm_list import PSMList
 
 
 class XTandemReader(ReaderBase):
-    def __init__(self, filename: Union[str, Path], decoy_prefix="DECOY_") -> None:
+    def __init__(
+        self, filename: str | Path, *args, decoy_prefix="DECOY_", **kwargs
+    ) -> None:
         """
         Reader for X!Tandem XML PSM files.
 
@@ -151,9 +152,7 @@ class XTandemReader(ReaderBase):
         peptide_entry = entry["protein"][0]["peptide"]
         psm = PSM(
             peptidoform=self._parse_peptidoform(peptide_entry, entry["z"]),
-            spectrum_id=entry["support"]["fragment ion mass spectrum"]["note"].split(
-                " "
-            )[0],
+            spectrum_id=entry["support"]["fragment ion mass spectrum"]["note"],
             is_decoy=entry["protein"][0]["label"].startswith(self.decoy_prefix),
             score=-np.log(peptide_entry["expect"]),
             precursor_mz=entry["mh"] - mass.nist_mass["H"][0][0],